tools/web_search_tool.py
from langchain.tools import BaseTool
from typing import Optional
import requests
from bs4 import BeautifulSoup
import json

class WebSearchTool(BaseTool):
    name = "web_search_tool"
    description = "Search the web for current information on a given topic"
    
    def _run(self, query: str) -> str:
        """
        Realiza búsquedas web para obtener información actualizada
        """
        try:
            # Simulación de búsqueda web (en producción usarías APIs reales)
            # Por ejemplo: Google Custom Search API, Bing Search API, etc.
            
            # Por ahora retornamos información simulada
            search_results = {
                "query": query,
                "results": [
                    {
                        "title": f"Information about {query}",
                        "snippet": f"Latest developments and insights on {query}",
                        "url": f"https://example.com/{query.replace(' ', '-')}",
                        "date": "2024-01-01"
                    }
                ]
            }
            
            return json.dumps(search_results, indent=2)
            
        except Exception as e:
            return f"Error performing web search: {str(e)}"
    
    async def _arun(self, query: str) -> str:
        """
        Versión asíncrona del método de búsqueda
        """
        return self._run(query)
